refactor : 플레이어 이동 로직 변경 및 스크립트 분리
