+++ unity/Assets/Prefabs/players/Lazy_cat.prefab
@@ -481,8 +481,8 @@ GameObject:
   m_Component:
   - component: {fileID: 4147768327354586}
   - component: {fileID: -1708393731892974340}
-  - component: {fileID: -1339487429236797604}
-  - component: {fileID: 8539331245908790726}
+  - component: {fileID: -284377965283096084}
+  - component: {fileID: -365612669801639028}
   m_Layer: 0
   m_Name: Lazy_cat
   m_TagString: Player
@@ -518,25 +518,13 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 76aadac887eb5704d98ba9ec59f6eb6e, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  animator: {fileID: 0}
-  SJump: {fileID: 8300000, guid: 2c5292ba5cf254919a6b9e2b18e04f49, type: 3}
   SDie: {fileID: 8300000, guid: 4b4dd7245f6104c8c8eb07ccf6b386ed, type: 3}
   SDamaged: {fileID: 8300000, guid: b9eff218e132a4e76be78840eca96fb2, type: 3}
   SRespawn: {fileID: 8300000, guid: 4e1ff8faed12a475eba9be0e46072269, type: 3}
-  SItemEnd: {fileID: 8300000, guid: a235d590ef2a04d1b85b61ef4c21337c, type: 3}
-  maxSpeed: 10
-  minspeed: 1
-  maxJumpPower: 10
-  minJumpPower: 0
-  currentSpeed: 0
-  currentJumpPower: 0
-  canJump: 1
-  tagOfFooting: 
-  velocity: 0
   _state: 0
   easyEnemyType: Easy
---- !u!65 &-1339487429236797604
-BoxCollider:
+--- !u!143 &-284377965283096084
+CharacterController:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
@@ -546,24 +534,32 @@ BoxCollider:
   m_IsTrigger: 0
   m_Enabled: 1
   serializedVersion: 2
-  m_Size: {x: 0.014, y: 0.02, z: 0.025}
-  m_Center: {x: 0, y: 0.009, z: 0}
---- !u!54 &8539331245908790726
-Rigidbody:
+  m_Height: 0
+  m_Radius: 0.02
+  m_SlopeLimit: 45
+  m_StepOffset: 0.3
+  m_SkinWidth: 0.08
+  m_MinMoveDistance: 0.001
+  m_Center: {x: 0, y: 0.017, z: 0.01}
+--- !u!114 &-365612669801639028
+MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1225627572341082}
-  serializedVersion: 2
-  m_Mass: 1
-  m_Drag: 0
-  m_AngularDrag: 0.05
-  m_UseGravity: 1
-  m_IsKinematic: 0
-  m_Interpolate: 0
-  m_Constraints: 0
-  m_CollisionDetection: 0
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: acbbfe4acb2eb704b81e389d221563cb, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  playerController: {fileID: -284377965283096084}
+  SJump: {fileID: 8300000, guid: 2c5292ba5cf254919a6b9e2b18e04f49, type: 3}
+  SItemEnd: {fileID: 8300000, guid: a235d590ef2a04d1b85b61ef4c21337c, type: 3}
+  moveDir: {x: 0, y: 0, z: 0}
+  <velocity>k__BackingField: 0
+  currentSpeed: 0
+  currentJumpPower: 0
 --- !u!1 &1242129020982672
 GameObject:
   m_ObjectHideFlags: 0
