+++ unity/Assets/Scripts/manager/LabelManager.cs
@@ -6,4 +6,6 @@ using System;
 public static class LabelManager
 {
     public const string DEV = "Development", PLAYGROUND = "Playground", SELECT_CHARACTER= "CharacterSelect";
+
+    public static string playScene = PLAYGROUND;
 }
