+++ unity/Assets/Scripts/screen/StageProgressBar.cs
