+++ unity/Assets/Scripts/screen/StageProgressBar.cs.meta
